Rocksd Dev Deployment

+++ libraries/chain/include/steem/chain/history_object.hpp
@@ -30,8 +30,8 @@ namespace steem { namespace chain {
          transaction_id_type  trx_id;
          uint32_t             block = 0;
          uint32_t             trx_in_block = 0;
-         uint16_t             op_in_trx = 0;
-         uint64_t             virtual_op = 0;
+         uint32_t             op_in_trx = 0;
+         uint32_t             virtual_op = 0;
          time_point_sec       timestamp;
          buffer_type          serialized_op;
