+++ libraries/plugins/apis/condenser_api/include/steem/plugins/condenser_api/condenser_api.hpp
@@ -87,8 +87,8 @@ struct api_operation_object
    transaction_id_type  trx_id;
    uint32_t             block = 0;
    uint32_t             trx_in_block = 0;
-   uint16_t             op_in_trx = 0;
-   uint64_t             virtual_op = 0;
+   uint32_t             op_in_trx = 0;
+   uint32_t             virtual_op = 0;
    fc::time_point_sec   timestamp;
    legacy_operation     op;
 };
